Fixed bug in resource_package.rs
When a rpk2 file is parsed it will use the patch_id variable to check if the package is a patch.
The parser used to ignore that and would try to parse anything the user had flagged as a patch. It now checks the user input AND the patch_id