Expose game mounting logic from the library
Some new utility functions have been added to the library, which allow users to mount games without having to deal with the low-level details of the game's folder structure and package definition files. This should reduce the amount of work needed by library users, and make sure that game mounting is done correctly and consistently.