Refactored the ResourcePackageBuilder
The ResourcePackageBuilder has enough information to determine the package filename by itself. It now do this when the user has called build with a directory and not a file path.

There is also no need to pass in the is_patch bool with a build anymore. We can just check this through a local param.

Adding a code example and replacing some parameters has also simplified the API a bit more.

For example: A user will now need to call the use_legacy_references() once to get legacy reference flags, instead of having to specify it with every build() call